+++ jdaviz/core/tools.py
@@ -850,6 +850,29 @@ class SelectFootprintOverlay(CheckableTool, HubListener):
 
 
 @viewer_tool
+class SkewerSelectRegion(CheckableTool, HubListener):
+    icon = os.path.join(ICON_DIR, 'skewer_select.svg')
+    tool_id = 'jdaviz:skewerregion'
+    action_text = 'Select/identify smallest region containing cursor'
+    tool_tip = 'Select/identify smallest region containing cursor'
+
+    def activate(self):
+        self.viewer.add_event_callback(self.on_mouse_event,
+                                       events=['click'])
+
+    def deactivate(self):
+        self.viewer.remove_event_callback(self.on_mouse_event)
+
+    def on_mouse_event(self, data):
+        msg = FootprintOverlayClickMessage(data, mode="skewer", sender=self)
+        self.viewer.session.hub.broadcast(msg)
+
+    def is_visible(self):
+        return any(isinstance(m, RegionOverlay) and m.visible
+                   for m in self.viewer.figure.marks)
+
+
+@viewer_tool
 class SelectRegionOverlay(CheckableTool, HubListener):
     icon = os.path.join(ICON_DIR, 'footprint_select.svg')
     tool_id = 'jdaviz:selectregion'
@@ -864,7 +887,7 @@ class SelectRegionOverlay(CheckableTool, HubListener):
         self.viewer.remove_event_callback(self.on_mouse_event)
 
     def on_mouse_event(self, data):
-        msg = FootprintOverlayClickMessage(data, sender=self)
+        msg = FootprintOverlayClickMessage(data, mode="nearest", sender=self)
         self.viewer.session.hub.broadcast(msg)
 
     def is_visible(self):